CLuaInstCurl: Fix compiler warning

+++ src/gui/lua/lua_curl.cpp
@@ -99,7 +99,7 @@ int CLuaInstCurl::CurlProgressFunc(void *p, curl_off_t dltotal, curl_off_t dlnow
 	curl_easy_getinfo(_pgd->curl, CURLINFO_RESPONSE_CODE, &responseCode);
 
 	uint32_t MUL = 0x7FFF;
-	uint32_t dlFragment = (dlnow * MUL) / dltotal;
+	uint32_t dlFragment = (uint32_t)((dlnow * MUL) / dltotal);
 	if (responseCode != 200) {
 		dlFragment = 0;
 		dlSpeed    = 0;
